requirements.txt
asgiref==3.8.1
Django==5.1.7
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.1
psycopg2-binary==2.9.10
sqlparse==0.5.3
tzdata==2025.1
django-cors-headers==4.7.0
setuptools==65.5.0
django-filter==23.2
dj-database-url==1.2.0
django-extensions==3.2.3
whitenoise==6.5.0
python-decouple==3.8
gunicorn==20.1.0
